Fix Tests

+++ self_serve_platform/tests/unit/test_data_extractor_pymupdf.py
@@ -186,6 +186,8 @@ def clear_image_folder(config):  # pylint: disable=W0621
     """
     if os.path.exists(config["image_output_folder"]):
         for file in os.listdir(config["image_output_folder"]):
+            if file == ".gitkeep":
+                continue  # Skip deleting the .gitkeep file
             file_path = os.path.join(config["image_output_folder"], file)
             if os.path.isfile(file_path):
                 os.remove(file_path)
@@ -270,7 +272,7 @@ def test_check_no_image_extraction(sample_pdf_path, config):  # pylint: disable=
     config["extract_image"] = False
     extractor_no_image = PyMuPdfForSectionsDataExtractor(config)
     extractor_no_image.parse(sample_pdf_path)
-    assert os.listdir(config["image_output_folder"]) == [], \
+    assert os.listdir(config["image_output_folder"]) == [".gitkeep",], \
         "Images were extracted even when flag is False"
